Add support for embedding PDF export of Google slideset
This requires a pending new release of i-slide https://github.com/w3c/i-slide/pull/47

The way this works is that when scribe.perl identifies the URL of a Google slideset in the slideset command, it downloads the PDF export of that slideset, embeds it as a data URL in a download link (which has the benefit of archiving it in the state it was at the time where the minutes are generated) and annotates i-slide elements so that i-slide knows how to display them from the embedded data.

Since downloading the PDF takes more time than what is usually expected for running scribe.perl, it would be good to ensure this doesn't freeze RRSAgent when it does so.